swf: Move helper functions for generating test data into `test_data`
